requirements_server.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
python-dotenv
